requirements.txt
asgiref==3.5.2
Django==4.1.1
django-exclusivebooleanfield==0.3.0
Pillow==9.2.0
python-decouple==3.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.2
